[Tilesheet] Rename logic nodes
