+++ armory/blender/arm/logicnode/animation/LN_play_tilesheet_action.py
@@ -1,16 +1,16 @@
 from arm.logicnode.arm_nodes import *
 
-class PlayTilesheetNode(ArmLogicTreeNode):
+class PlayTilesheetActionNode(ArmLogicTreeNode):
     """Plays the given tilesheet action."""
-    bl_idname = 'LNPlayTilesheetNode'
-    bl_label = 'Play Tilesheet'
+    bl_idname = 'LNPlayTilesheetActionNode'
+    bl_label = 'Play Tilesheet Action'
     arm_version = 1
     arm_section = 'tilesheet'
 
     def arm_init(self, context):
         self.add_input('ArmNodeSocketAction', 'In')
         self.add_input('ArmNodeSocketObject', 'Object')
-        self.add_input('ArmStringSocket', 'Name')
+        self.add_input('ArmStringSocket', 'Action')
 
         self.add_output('ArmNodeSocketAction', 'Out')
         self.add_output('ArmNodeSocketAction', 'Done')
